fold sglang worker group helper
