+++ components/src/dynamo/sglang/_disagg.py
@@ -4,13 +4,15 @@
 """Disaggregation helpers shared by the legacy (`init_llm.py`,
 `register.py`) and unified (`llm_engine.py`) entry points.
 
-Two helpers:
-
 * :func:`compute_bootstrap_address` — resolve the `(host, port)` triple a
   prefill worker advertises to decode peers from a live `sgl.Engine`.
   Returns `(None, None)` on any failure so legacy callers can keep their
   graceful-degradation behaviour; the unified path treats `None` as a
   fatal configuration error and raises.
+
+* :func:`get_sglang_worker_group_id` — normalize the SGLang multinode
+  `dist_init_addr` into a runtime-data key that non-leader workers can use
+  to discover the leader endpoint.
 
 * :func:`warmup_prefill_engine` — drive one request through the disagg
   path with SGLang's `FAKE_BOOTSTRAP_HOST` so the first real request
@@ -28,6 +30,16 @@ import sglang as sgl
 # Matches the prior in-tree value. Long enough for the slowest cold-start
 # we've seen (TP8 70B with FlashInfer JIT).
 _PREFILL_WARMUP_TIMEOUT_S = 1800.0
+
+SGLANG_WORKER_GROUP_ID_KEY = "sglang_worker_group_id"
+
+
+def _network_address_cls():
+    # Deferred to function body so pre-commit test collection can import
+    # `_disagg` without the SGLang network helper available.
+    from sglang.srt.utils.network import NetworkAddress
+
+    return NetworkAddress
 
 
 def compute_bootstrap_address(
@@ -81,6 +93,33 @@ def compute_bootstrap_address(
     except Exception as e:
         logging.warning("Failed to compute bootstrap address: %s", e)
         return None, None
+
+
+def get_sglang_worker_group_id(server_args) -> Optional[str]:
+    """Return the shared SGLang multi-node worker group id."""
+    nnodes = getattr(server_args, "nnodes", 1) or 1
+    if nnodes <= 1:
+        return None
+
+    dist_init_addr = getattr(server_args, "dist_init_addr", None)
+    if not dist_init_addr:
+        logging.warning(
+            "SGLang multi-node worker group id requires dist_init_addr; "
+            "falling back to local worker id behavior"
+        )
+        return None
+
+    try:
+        parsed = _network_address_cls().parse(str(dist_init_addr).strip())
+        resolved = parsed.resolved()
+        return f"dist_init:{resolved.to_tcp()}"
+    except Exception as e:
+        logging.warning(
+            "Failed to normalize SGLang dist_init_addr=%r for worker group id: %s",
+            dist_init_addr,
+            e,
+        )
+        return None
 
 
 async def warmup_prefill_engine(engine: sgl.Engine, bootstrap_port: int) -> None:
