+++ components/src/dynamo/sglang/register.py
@@ -23,11 +23,11 @@ from dynamo.llm import (
     register_model,
 )
 from dynamo.sglang._compat import get_scheduler_info
-from dynamo.sglang.args import DynamoConfig
-from dynamo.sglang.worker_group import (
+from dynamo.sglang._disagg import (
     SGLANG_WORKER_GROUP_ID_KEY,
     get_sglang_worker_group_id,
 )
+from dynamo.sglang.args import DynamoConfig
 
 SGLANG_HICACHE_MOONCAKE_RUNTIME_KEY = "sglang_hicache_mooncake"
